requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
bunch==1.0.1
cachetools==5.2.0
celery==5.2.7
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
configparser==5.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
cssselect==1.1.0
cssutils==2.6.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1
django-cors-headers==3.13.0
django-debug-toolbar==3.6.0
django-filter==22.1
django-render-block==0.9.1
django-templated-email==3.0.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docopt==0.6.2
drf-nested-routers==0.93.4
flower==1.2.0
humanize==4.4.0
idna==3.4
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.1
MarkupSafe==2.1.1
mock==4.0.3
mysqlclient==2.1.1
oauthlib==3.2.1
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
premailer==3.10.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
py==1.11.0
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
pytest-watch==4.2.0
python3-openid==3.2.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
simple-configparser==0.1.4.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
templated-mail==0.4.0.2
tomli==2.0.1
tornado==6.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
watchdog==2.1.9
wcwidth==0.2.5
wrapt==1.14.1
yagmail==0.15.293
